delete error fixed

+++ src/Pages/InventoryPage/InventoryPage.js
@@ -4,18 +4,21 @@ import InventoryTable from './InventoryTable';
 
 const InventoryPage = () => {
     const [books, setBooks] = Usebooks();
-   
+
     const handleDelete = (id) => {
-        const url = `http://localhost:5000/inventory/${id}`;
-        fetch(url, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-                const remaining = books.filter(book => book._id !== id);
-                setBooks(remaining);
-            })   
+        const proceed = window.confirm("Do you want to delete this item?");
+        if (proceed) {
+            const url = `http://localhost:5000/inventory/${id}`;
+            fetch(url, {
+                method: 'DELETE',
+            })
+                .then(res => res.json())
+                .then(data => {
+                    console.log(data);
+                    const remaining = books.filter(book => book._id !== id);
+                    setBooks(remaining);
+                })
+        }
     }
     return (
         <div className="container my-5">
@@ -33,7 +36,7 @@ const InventoryPage = () => {
                 </thead>
                 <tbody>
                     {
-                        books.map(book => <InventoryTable handleDelete={()=>handleDelete(book._id)} key={book._id} book={book}></InventoryTable>)
+                        books.map(book => <InventoryTable handleDelete={() => handleDelete(book._id)} key={book._id} book={book}></InventoryTable>)
                     }
                 </tbody>
             </table>
